MAJOR UPDATE: implementing cubic splines for input functions, see Schelker et al., Bioninformatics 2012. :-) For example usage, see examples/Swameye_PNAS2003.
For now splines with 3,4 or 5 knots are implemented. That should suffice for most cases, however, one can easily track the changes form this update am implement larger splines.
There are two basic types of splines implemented: normal splines and splines that are only positive.
Also, in addition to the knot points, the slope of the spline in the first knot can be given.