Fix bug in arInitHierarchical
The scale parameters suitable for hierarchical optimization are those
which appear _only_ in observables of form `s*xz` where `s` is the
subject scale parameter and `xz` is a model species or a derived
variable. Until now, arInitHierarchical has been detecting parameters
which appear _at least in one_ observable of form `s*xz` (meaning that
they were allowed to appear also in obervables of other form).
Consequently, parameters not suitable for the hierarchical optimization
could be detected by arInitHierarchical and the results of the
hierarchical optimization could be not correct. This commit fixes it by
adding extra checks in arInitHierarchical to make the latter function
detect only the right parameters.