Add necessary assertion checks in arCalcHierarchical
The implementation of hierarchical optimization depends on the specific
form of the merit function. We currently assume the following one:

    J(s,p) = \sum_j (data_j - s*x(p,t_j))^2/std_j^2

Each feature that modifies this expression also breaks the method of
hierarchical optimization, making its results incorrect. To avoid this,
in arCalcHierarchical we try to check if d2d features affecting the form
of the merit function are enabled. This commit adds extra checks of this
kind in arCalcHierarchical, hopefully making the checklist complete (or
close to complete).