Test feasibility of hierarchical optimization
Hierarchical optimization is not feasible if the solution of the model
is zero for all measurement times. This commit adds the respective
assertion in arCalchierarchical.

This extra assertion is not strictly necessary. Degeneration of the
solution to zero would result in NaN scales, subsequently resulting in
NaN residuals, which would trigger an error anyway. Nevertheless, with
the extra assertion we get a more specific error message which may make
it easier to debug.