Fix a bug in assertions in arCalcHierarchical
Currently arCalcHierarchical requires no dependence of errors on any
fittable parameters, neither explicitly nor implicitly. Until now,
arCalcHierarchical has been asserting that there is no fitting of error
parameters, but at the same time has been allowing implicit dependence
of errors on fittable parameters of the model via observables. This is
a bug since with such implicit dependence of errors on fittable
parameters the results of arCalcHierarchical are incorrect. This commit
fixes it by asserting the use of experimental errors, which are
independent of the fittable parameters by definition.