Fix a bug in arCalcHierarchical
For certain purposes, in arCalcHierarchical we have been testing if all
elements ar.model(im).data(id).useHierarchical are NaNs. But this cannot
happen since the array useHierarchical is constructed such that its
elements may be either false or true (and not NaN). So we should rather
test if the elements of useHierarchical are all false. This commit fixes
this issue.