Minor change: convert to symbolic in MATLAB2020a
It seems MATLAB2020a cannot convert some char strings to symbolic arrays. Apparently, the problem is because the expression which is going to be converted is a column vector. By replacing it with a row vector, MATLAB can handle the conversion. What I suggest is explicitly replacing semicolons with commas in MATLAB2020a and later versions.